feat: serve the page verbatim and move its data to a JSON API
GET / now streams the embedded page straight from flash with send_P and
no substitution at all, so the bytes never exist in the heap. That
removes the copy per request plus one reallocation per placeholder, which
was the cost that made growing the page unaffordable. The response
carries an ETag, so the second visit costs a 304 instead of the whole
page over the 4G link.

The ETag hashes the page content rather than reusing the firmware version
string. The two do not move together during development: editing the HTML
without a new commit leaves the version identical, and the browser would
keep serving the stale page with the symptom "my change did not show up".

New routes: GET /api/status for everything that changes on its own, GET
/api/config for what the form edits. Passwords are absent from the config
response, not masked — a mask travels back on the next POST and would
become the new password, and the length alone already says more than it
should.

POST /api/config keeps accepting urlencoded bodies instead of JSON. The
WebServer already parses that into server_.arg(), while JSON would mean
hand-writing a parser in firmware — nesting, escapes, unicode — for eight
flat fields the browser encodes with URLSearchParams in one line. This
deviates from what PRD 12 states; the PRD is updated in the same change.
Responses are JSON everywhere, including /update and /firmware/confirmar,
so the page never has to guess the body type from the status code.

/firmware/confirmar answers 200 with JSON instead of redirecting to /. The
303 existed to keep F5 from resubmitting the POST, and there is no
navigation left to reload; meanwhile fetch() would follow the redirect and
download the entire page only to discard it.

collectHeaders is called after server_.begin() because begin() resets the
list. Without it If-None-Match arrives empty, the 304 never happens, and
the full page ships on every visit with nobody noticing.

The battery reading comes from a provider holding the value the loop
already measured, not from a fresh readVoltage(). That call blocks about
100 ms averaging out ADC noise, and 100 ms inside the handler would delay
the response and everything after handleClient() on every status poll.

escapeForHtmlAttribute is gone with the last replace, and html_page.cpp
with it. The only escaping left is JSON, in domain/, where it is testable
without a board.

+++ src/domain/string_type.h
@@ -38,7 +38,9 @@ inline String numberToString(uint32_t value) {
 // bateria sairia diferente na placa e no teste. Quem chama diz quantas casas quer.
 inline String numberToString(float value, uint8_t decimals) {
 #ifdef ARDUINO
-  return String(value, decimals);
+  // O cast nao e enfeite: com uint8_t a chamada fica ambigua entre String(float, unsigned
+  // int) e String(long long, unsigned char) — nenhuma das duas vence nos dois argumentos.
+  return String(value, static_cast<unsigned int>(decimals));
 #else
   char buffer[32];
   std::snprintf(buffer, sizeof(buffer), "%.*f", static_cast<int>(decimals), value);
